Model instance and class attributes in class field
Summary:
In this diff, we eliminate ClassFieldInner::Simple in favor of
ClassFieldInner::{Class,Instance}Attribute

This completes the core refactor toward modeling class field semantics in an
enum. THe diff is a bit big but it is easier to make the change and have
conffidence if we do this in one shot because the compiler can verify we handle
all cases.

Reviewed By: grievejia

Differential Revision: D88285161

fbshipit-source-id: 86a55e2165dae8c297935c23bb57ce258a5bfea6