Deduplicate scope stack traversals
Summary:
It was noted in review of D88405055 that the `look_up_name_for_read` and
`suggest_similar_name` functions share quite a bit of fiddly logic related to
traversing the scope stack and handling class body scope correctly.

By pulling that plumbing into a (private) helper and using callbacks for the
divergent parts of the logic, we can avoid duplication. The resulting code is
a bit harder to skim, but this is probably worthwhile to avoid brittleness
maintaining two copies of the fiddly traversal logic.

Reviewed By: rchen152

Differential Revision: D88426054

fbshipit-source-id: 7d020d711fba0bd8a75e26de466b333a18b49e92