check for is_subset_eq(Never) instead of is_any
Summary:
Normally if a class extends `Any` it should behave sort of like `Any`, in the sense that it's assignable everywhere.

However, there are places in the codebase where we say "error if this value is assignable to _" and "error if this value is NOT assignable to _".

IMO, if we have `Any`, we should do whatever causes it to not error.

Since both cases use the `unwrap_X` helpers, how we behave depends on the control flow at the call site, meaning there is no way to build the check into the helper. Each call site will need to remember to check, which makes these helpers pretty dangerous to use.

`Type::is_any` is not enough, we need to make sure it's not a class that extends Any, or a type var bounded by a class that extends Any, etc...

The solution I came up with is to check if it's assignable to `Never`

Context and broader motivations: Context: https://github.com/facebook/pyrefly/issues/1624#issuecomment-3617675898

fixes https://github.com/facebook/pyrefly/issues/1624 by addressing item 1

We should still go and address 2 and 3 later though

Reviewed By: stroxler

Differential Revision: D88530042

fbshipit-source-id: 18b45a9e3296c452b07f5d570888bba71e78612a