Remove Arc from HeavyTaskQueue
Summary:
Now that we can borrow in the server threads, we don't need an Arc here.

As with the previous diff, the title is a bit misleading, because anything closed over in a HeavyTask itself needs to be Arc due to the `'static` lifetime. Currently the only such HeavyTaskQueue is sourcedb_queue.

Reviewed By: kinto0

Differential Revision: D88534254

fbshipit-source-id: 1346a88727251d3da730d66e3db0f67c0f61cc6c