Stop listening for LSP requests after Exit notification
Summary:
Before this diff, dispatch_lsp_events would not return after receiving an Exit notification, meaning the thread running this loop would not exit.

In the next diff, I move the server threads into a thread scope, but all threads in a scope need to exit before the scope ends. This fix ensures that the recv thread can be joined when we receive an exit notification.

Reviewed By: ndmitchell

Differential Revision: D88507725

fbshipit-source-id: 6e8d47c3affcf11370d2c4a84705b3984d8ef3cf