Run server threads in a thread scope
Summary:
The LSP server uses a handful of threads:

1. recv thread: receives LSP client messages and dispatches to main thread, useful for looking ahead to see canceled requests
2. recheck thread: performs committing transactions in response to invalidations changes, triggers diagnostics on main thread
3. references thread: performs potentially long-running, non-committable work. individual jobs can be canceled mid-way.
4. sourcedb thread: deals with build systems like Buck2, where we need to look at build output.

Before this diff, we spawn these threads and never join them. To the Rust borrow checker, this means the thread may outlive anything, so we can't capture anything that might have a shorter lifespan. To work around this, we just wrap everything we need in Arc and move them into the thread.

In practice, these threads have the same lifetime as the server itself. It would be very convenient to just borrow things from the server in these threads.

This diff moves the theads into a scope, which gives them a real lifetime. Doing this will make it possible to borrow and we can eliminate a bunch of unnecessary Arcs.

To be clear, the motivation is clarity. Creating a handful of Arcs is not a performance issue.

Reviewed By: stroxler

Differential Revision: D88514484

fbshipit-source-id: cf87515ff6c58ac7bff3ac0e64ced7e627c966c7