Prepare HeavyTaskQueue for defunctionalization
Summary:
Currently the heavy task queue stores closures, and these closures are stored with a `'static` lifetime, which is somewhat awkward to deal with.

I'd like to stop storing closures in the task queue. Defunctionalizing means we will store data instead of code in the queue, representing what action we want to take.

This diff explicitly parameterizes HeavyTaskQueue with HeavyTask, and moves the responsibility of "running" the heavy task to the individual thread. This should be a behavior-preserving change.

Reviewed By: yangdanny97

Differential Revision: D88609489

fbshipit-source-id: 6e84b0301f593917460225fb3088ef93b2263d2b