Add create_archive helper function to build.rs
Summary:
Add a reusable `create_archive` function that creates a compressed tar archive from a given input path and writes it along with a SHA256 digest. This function will be used in a follow-up commit to support bundling additional archives (e.g., third-party stubs not from typeshed).

The function is currently marked as `#[allow(dead_code)]` since it's not yet called from main.

Reviewed By: yangdanny97

Differential Revision: D88486624

fbshipit-source-id: c99600239d0741ceba95230eadfa1ceeadb84e3d