Refactor main function in fbcode/pyrefly/crates/pyrefly_bundled/build.rs to use create_archive.
Summary: In order to prepare the bundled crate to work with other third party stubs besides typeshed, we need to make the logic here more general. This diff uses the new create_archive function in order to bundled the existing typeshed stubs. In a subsequent diff we will also use this function for the other third party stubs.

Reviewed By: yangdanny97

Differential Revision: D88486628

fbshipit-source-id: 36d9bc5353d5ee7add15e83e0b7d9e15129d8909