Update find_third_party_typeshed_stub to also find other third party stubs.
Summary:
Currently we have the function find_third_party_typeshed_stub which looks to see if we have bundled typeshed stubs for a particular module. Now we have also expanded this to look for non-typeshed stubs as well.

We do make a decision here to prioritize typeshed stubs over other bundled stubs. Since typeshed for now is updated more regularly, if we do have a stub from typeshed it might make more sense to use that one. If we do find a typeshed stub for a particular module, we will immediately return it rather than also checking for a non-typeshed third party stub.

Reviewed By: kinto0

Differential Revision: D88486679

fbshipit-source-id: f39d051b42efbb97b18c67441effa535f64b5994