Better representation for canonical types
Summary:
# Problem
Pysa support modeling a class attribute (NOT an instance attribute) as a source or a sink.
However, when running Pysa from Pyrefly's type information, we don't have enough information from types to know if an expression is a class name (i.e `type[A]`)

# Solution
Our current solution is to export a `ClassNamesFromType` type that represents a list of classes with a set of flag modifiers (optional, readonly, etc.). We could potentially add a `type: bool` modifier, but it would be ambiguous, because an expression can have a mix: `type[A] | B`.

This diff proposes to update the representation to be `ClassWithModifiers.t list` where `ClassWithModifiers.t = { class_name: string; modifiers = Modifier.t list }`. This allows to represent types more accurately.

In the next diff, we will solve the actual problem with class attributes.

Reviewed By: tianhan0

Differential Revision: D88659711

fbshipit-source-id: 9797cc718afe56dbfe1562f35ca1685936d2c6c9