Use aliases for atomic types only
Summary:
**Context**:

I am implementing better types for lax mode for pydantic, which is a feature requested by a user (https://github.com/facebook/pyrefly/issues/1538)

I had a discussion with Danny about the possibility of using named unions for pydantic lax mode because the signatures are getting too big.

The idea would be that for every type `T`, we have `LaxT` which represents the union of the acceptable types with `T` as an input.

After the discussion, the main challanges are:

- How should merging two named unions look like? we may need to consider a normal form

- is LaxT a good name/how will the user understand what it means?

A potential solution we discussed is to use aliasing instead of named unions. We would have to bundle those aliases with pyrefly. We would also need to look into what the error messages would look like with aliases and how much we have to add pydantic specific stuff to the typechecker.

I talked to Jia briefly as well, who suggested it can be a good idea to continue implementing the feature and then check if the users face any readibility issues.

So I thought that a good in between solution is to use named unions just for atomic types, not containers and then have clear documentation for this when we roll this out. We can then continue to look for better alternatives in case a better solution exists.

This will
- Be simple to read as it just applies to atomic types and we will add it to the documentation to make it very clear what the type stands for
- Make the signatures look better than they currently do till we complete the implementation
- give the users the new feature while we research other options

Since the feature is already implemented for some types, users will already start to see those large signatures. My suggestion is to add this and continue evaluating the other options. In the meanwhile, users will not experience readability issues. Other options include
- do nothing and keep implementing the feature and respond to feedback
- roll the feature back till we decide (but this is a useful feature so I think we should keep it)

What to people think?

cc yangdanny97 grievejia kinto0 (for opinions about readibility) jvansch1 (for opinions about bundling)

Reviewed By: yangdanny97, grievejia

Differential Revision: D88682121

fbshipit-source-id: 3e99f5fb91f46e810156b0f18efe42fcba3aaa4b